added timestamp to its own module and sorted my fetch calls
