+++ src/scripts/timestamp.js
@@ -1,0 +1,12 @@
+const timestamp = () => {
+  let currentDate = new Date()
+    let date = currentDate.getDate()
+    let month = currentDate.getMonth()
+    let year = currentDate.getFullYear()
+    let hour = currentDate.getHours()
+    let min = ('0' + currentDate.getMinutes()).slice(-2)
+    console.log(currentDate)
+    return `${month+1}-${date}-${year} ${hour}:${min}`
+}
+
+export default timestamp
